Comparison of narrow type with wide type in loop condition

+++ Architecture/amd64/cpu/SymmetricMultiprocessing.cpp
@@ -84,7 +84,7 @@ namespace SMP
             KPrint("VirtualBox detected, disabling SMP");
             return;
         }
-        for (uint8_t i = 0; i < ((ACPI::MADT *)madt)->CPUCores + 1; i++)
+        for (uint16_t i = 0; i < ((ACPI::MADT *)madt)->CPUCores + 1; i++)
         {
             debug("Initializing CPU %d", i);
             if ((((APIC::APIC *)Interrupts::apic[0])->Read(APIC::APIC::APIC_ID) >> 24) != ((ACPI::MADT *)madt)->lapic[i]->ACPIProcessorId)
